refactor: 전략패턴 도입

+++ src/main/java/janggi/piece/pieces/Piece.java
@@ -15,7 +15,7 @@ public interface Piece {
 
     PieceType getType();
 
-    int getScore();
+    double getScore();
 
     Team getTeam();
 }
